[ 포스팅 페이지 DI에러 수정 및 Dao return type  수정 ] - di는 rxjava를 디펜던시 추가하라고 해서 함 - 다오 @Insert는 리턴타입 Long

+++ data/src/main/java/com/namu/data/dao/PostDao.kt
@@ -10,7 +10,7 @@ import io.reactivex.Single
 @Dao
 interface PostDao {
      @Insert(onConflict = OnConflictStrategy.REPLACE)
-     fun savePost(item:LocalDataPostModel): Single<Int>
+     fun savePost(item:LocalDataPostModel): Single<Long>
 
      @Query("SELECT * FROM localdatapostmodel WHERE id =:id")
      fun getPostEach(id: Int):  Single<LocalDataPostModel>
